Add descriptors and USB
Not actually working yet, but it does enumerate properly and send
descriptors to the host.

My descriptors are almost identical to the keyboard demo, so I just have
the Makefile copy Descriptors.* from the demo and tweak it with sed.
This causes some minor issues with the LUFA build system.